fixed loading issue with textarea and added eventListener to get input from users
