+++ app/javascript/application.js
@@ -4,4 +4,3 @@ import "./controllers"
 import "bootstrap"
 import "./codemirror/codemirror"
 import "./codemirror/ruby"
-import "./codemirror/text_area"
